requirements.txt
google-genai==1.11.0
requests==2.32.3
Appium-Python-Client==1.3.0